Add maximum deflate ratio to MslCompression and CompressionImpl. Default is a factor of 200. Rename and move lzw.js to LzwCompression.js.
Fix a major LZW uncompression bug in JavaScript implementation that would treat values of 0 (zero) as false instead of a valid translation.
