+++ core/src/main/java/com/netflix/msl/util/MslCompression.java
@@ -1,6 +1,6 @@
 /**
- * Copyright (c) 2017 Netflix, Inc.  All rights reserved.
- * 
+ * Copyright (c) 2017-2018 Netflix, Inc.  All rights reserved.
+ *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
@@ -32,7 +32,7 @@ import com.netflix.msl.io.LZWOutputStream;
 /**
  * <p>Data compression and uncompression. Can be configured with a backing
  * implementation.</p>
- * 
+ *
  * <p>This class is thread-safe.</p>
  *
  * @author Wesley Miaw <wmiaw@netflix.com>
@@ -40,31 +40,40 @@ import com.netflix.msl.io.LZWOutputStream;
 public class MslCompression {
     /** Registered compression implementations. */
     private static Map<CompressionAlgorithm,CompressionImpl> impls = new ConcurrentHashMap<CompressionAlgorithm,CompressionImpl>();
-    
+    /** Maximum deflate ratio. Volatile should be good enough. */
+    private static volatile int maxDeflateRatio = 200;
+
     /**
      * <p>A data compression implementation. Implementations must be thread-
      * safe.</p>
      */
     public static interface CompressionImpl {
         /**
-         * Compress the provided data.
-         * 
+         * <p>Compress the provided data.</p>
+         *
          * @param data the data to compress.
-         * @return the compressed data.
+         * @return the compressed data. May also return {@code null} if the
+         *         compressed data would exceed the original data size.
          * @throws IOException if there is an error compressing the data.
          */
         public byte[] compress(final byte[] data) throws IOException;
-        
+
         /**
-         * Uncompress the provided data.
-         * 
+         * <p>Uncompress the provided data.</p>
+         *
+         * <p>If the uncompressed data ever exceeds the maximum deflate ratio
+         * then uncompression must abort and an exception thrown.</p>
+         *
          * @param data the data to uncompress.
+         * @param maxDeflateRatio the maximum deflate ratio.
          * @return the uncompressed data.
-         * @throws IOException if there is an error uncompressing the data.
+         * @throws IOException if there is an error uncompressing the data or
+         *         if the ratio of uncompressed data to the compressed data
+         *         ever exceeds the specified deflate ratio.
          */
-        public byte[] uncompress(final byte[] data) throws IOException;
+        public byte[] uncompress(final byte[] data, final int maxDeflateRatio) throws IOException;
     }
-    
+
     /**
      * Default GZIP compression implementation.
      */
@@ -83,20 +92,27 @@ public class MslCompression {
             }
             return baos.toByteArray();
         }
-        
+
         /* (non-Javadoc)
-         * @see com.netflix.msl.util.MslCompression.CompressionImpl#uncompress(byte[])
+         * @see com.netflix.msl.util.MslCompression.CompressionImpl#uncompress(byte[], int)
          */
         @Override
-        public byte[] uncompress(final byte[] data) throws IOException {
+        public byte[] uncompress(final byte[] data, final int maxDeflateRatio) throws IOException {
             final ByteArrayInputStream bais = new ByteArrayInputStream(data);
             final GZIPInputStream gzis = new GZIPInputStream(bais);
             try {
                 final byte[] buffer = new byte[data.length];
                 final ByteArrayOutputStream baos = new ByteArrayOutputStream(data.length);
                 while (buffer.length > 0) {
+                    // Uncompress.
                     final int bytesRead = gzis.read(buffer);
                     if (bytesRead == -1) break;
+
+                    // Check if the deflate ratio has been exceeded.
+                    if (baos.size() + bytesRead > maxDeflateRatio * data.length)
+                        throw new IOException("Deflate ratio " + maxDeflateRatio + " exceeded. Aborting uncompression.");
+
+                    // Save the uncompressed data for return.
                     baos.write(buffer, 0, bytesRead);
                 }
                 return baos.toByteArray();
@@ -105,7 +121,7 @@ public class MslCompression {
             }
         }
     }
-    
+
     /**
      * Default LZW compression implementation.
      */
@@ -124,20 +140,27 @@ public class MslCompression {
             }
             return baos.toByteArray();
         }
-        
+
         /* (non-Javadoc)
-         * @see com.netflix.msl.util.MslCompression.CompressionImpl#uncompress(byte[])
+         * @see com.netflix.msl.util.MslCompression.CompressionImpl#uncompress(byte[], int)
          */
         @Override
-        public byte[] uncompress(final byte[] data) throws IOException {
+        public byte[] uncompress(final byte[] data, final int maxDeflateRatio) throws IOException {
             final ByteArrayInputStream bais = new ByteArrayInputStream(data);
             final LZWInputStream lzwis = new LZWInputStream(bais);
             try {
                 final byte[] buffer = new byte[data.length];
                 final ByteArrayOutputStream baos = new ByteArrayOutputStream(data.length);
                 while (buffer.length > 0) {
+                    // Uncompress.
                     final int bytesRead = lzwis.read(buffer);
                     if (bytesRead == -1) break;
+
+                    // Check if the deflate ratio has been exceeded.
+                    if (baos.size() + bytesRead > maxDeflateRatio * data.length)
+                        throw new IOException("Deflate ratio " + maxDeflateRatio + " exceeded. Aborting uncompression.");
+
+                    // Save the uncompressed data for return.
                     baos.write(buffer, 0, bytesRead);
                 }
                 return baos.toByteArray();
@@ -146,16 +169,16 @@ public class MslCompression {
             }
         }
     }
-    
+
     static {
         MslCompression.register(CompressionAlgorithm.GZIP, new GzipCompressionImpl());
         MslCompression.register(CompressionAlgorithm.LZW, new LzwCompressionImpl());
     }
-    
+
     /**
      * <p>Register a compression algorithm implementation. Pass {@code null} to
      * remove an implementation.</p>
-     * 
+     *
      * @param algo the compression algorithm.
      * @param impl the data compression implementation. May be {@code null}.
      */
@@ -167,8 +190,22 @@ public class MslCompression {
     }
 
     /**
+     * <p>Sets the maximum deflate ratio used during uncompression. If the
+     * ratio is exceeded uncompression will abort.</p>
+     *
+     * @param deflateRatio the maximum deflate ratio.
+     * @throws IllegalArgumentException if the specified ratio is less than
+     *         one.
+     */
+    public static void setMaxDeflateRatio(final int deflateRatio) {
+        if (deflateRatio < 1)
+            throw new IllegalArgumentException("The maximum deflate ratio must be at least one.");
+        MslCompression.maxDeflateRatio = deflateRatio;
+    }
+
+    /**
      * Compress the provided data using the specified compression algorithm.
-     * 
+     *
      * @param compressionAlgo the compression algorithm.
      * @param data the data to compress.
      * @return the compressed data or null if the compressed data would be larger than the
@@ -181,15 +218,15 @@ public class MslCompression {
             throw new MslException(MslError.UNSUPPORTED_COMPRESSION, compressionAlgo.name());
         try {
             final byte[] compressed = impl.compress(data);
-            return (compressed.length < data.length) ? compressed : null;
+            return (compressed != null && compressed.length < data.length) ? compressed : null;
         } catch (final IOException e) {
             throw new MslException(MslError.COMPRESSION_ERROR, "algo " + compressionAlgo.name(), e);
         }
     }
-    
+
     /**
      * Uncompress the provided data using the specified compression algorithm.
-     * 
+     *
      * @param compressionAlgo the compression algorithm.
      * @param data the data to uncompress.
      * @return the uncompressed data.
@@ -200,7 +237,7 @@ public class MslCompression {
         if (impl == null)
             throw new MslException(MslError.UNSUPPORTED_COMPRESSION, compressionAlgo.name());
         try {
-            return impl.uncompress(data);
+            return impl.uncompress(data, maxDeflateRatio);
         } catch (final IOException e) {
             throw new MslException(MslError.UNCOMPRESSION_ERROR, "algo " + compressionAlgo.name(), e);
         }
