+++ tests/src/test/javascript/msltests.html
@@ -20,7 +20,6 @@
 
 <!-- libraries -->
 <script src="../../../../core/src/main/javascript/lib/asnjwk.js"></script>
-<script src="../../../../core/src/main/javascript/lib/lzw.js"></script>
 <script src="../../../../core/src/main/javascript/lib/promise.js"></script>
 
 <!-- source files -->
@@ -219,10 +218,10 @@
 <script type="text/javascript" src="../../main/javascript/userauth/MockUserIdTokenAuthenticationFactory.js"></script>
 
 <!-- test specifications -->
-<script type="text/javascript" src="lib/lzwtest.js"></script>
 <script type="text/javascript" src="util/ArraysTest.js"></script>
 <script type="text/javascript" src="util/Base64Test.js"></script>
 <script type="text/javascript" src="util/ClassTest.js"></script>
+<script type="text/javascript" src="util/LzwCompressionTest.js"></script>
 <script type="text/javascript" src="util/RandomTest.js"></script>
 <script type="text/javascript" src="util/TextEncodingTest.js"></script>
 <script type="text/javascript" src="MslConstantsTest.js"></script>
